refactor: fixed fxml to work with different locs
